+++ app/decorators/dc20_character/base_decorator.rb
@@ -2,13 +2,16 @@
 
 module Dc20Character
   class BaseDecorator < SimpleDelegator
-    delegate :id, :name, :data, to: :__getobj__
+    delegate :id, :name, :data, :feats, to: :__getobj__
     delegate :abilities, :main_class, :level, :combat_expertise, :health, :classes, :attribute_points, :ancestries, :skill_levels,
              :skill_expertise, :skill_points, :skill_expertise_points, :trade_points, :trade_expertise_points, :language_points,
              :trade_levels, :trade_expertise, :trade_knowledge, :language_levels, :conditions, to: :data
 
     def parent = __getobj__
-    def method_missing(_method, *args); end
+
+    def method_missing(method, *args) # rubocop: disable Lint/UnusedMethodArgument
+      __getobj__.respond_to?(method.to_sym) ? __getobj__.public_send(method) : nil
+    end
 
     def modified_abilities
       @modified_abilities ||= abilities.merge('prime' => abilities.values.max)