+++ app/decorators/pathfinder2_character/base_decorator.rb
@@ -5,7 +5,7 @@ module Pathfinder2Character
     FLEXIBLE_SKILLS = %w[acrobatics athletics].freeze
     ARMOR_ABILITIES = %w[str dex].freeze
 
-    delegate :id, :name, :data, to: :__getobj__
+    delegate :id, :name, :data, :feats, to: :__getobj__
     delegate :race, :subrace, :main_class, :classes, :subclasses, :level, :languages, :health, :selected_skills,
              :lore_skills, :background, :weapon_skills, :armor_skills, :main_ability, :class_dc, :conditions,
              :saving_throws, :dying_condition_value, :ability_boosts, :skill_boosts, :coins, to: :data