+++ app/serializers/pathfinder2/character_serializer.rb
@@ -22,7 +22,7 @@ module Pathfinder2
     def decorator
       @decorator ||= {}
       @decorator.fetch(object.id) do |key|
-        @decorator[key] = object.decorator
+        @decorator[key] = object.decorator(simple: (context ? (context[:simple] || false) : false))
       end
     end
   end